Don't use this version
The makefile deletes the header file, I think it is because it is not
precious and make thinks it is safe to delete it (which is not). I'm not sure
why it does so, but I think I know where to find the answere. I push to git
however so that I can recover the latest version of the header file in case
I screw up which is probable.