Changed how the tests are written
I changed how the tests are written. Previously the new_test command
took 2 arguments, the first one being the name of the test and the
second one were the test body, the code executed in the test, required
to be a lambda function. I found this to not agreeing well with many editors,
including vim. What is worse is that I think that begginer programmers
might be hessitent to use lambda functions. Therefore, I changed it.

In the new format, the new_test command takes one argument, the name of
the test. The test body is then written as normal function body block.